+++ automation/services/watchdog/metrics.py
@@ -22,7 +22,7 @@ def collect_cluster_crashes(v1, namespace, cluster_crashes):
   pods = v1.list_namespaced_pod(namespace, watch=False)
 
   containers = list(itertools.chain(*[ pod.to_dict()['status']['container_statuses'] for pod in pods.items ]))
-  mina_containers = list(filter(lambda c: c['name'] in [ 'coda', 'seed', 'coordinator' ], containers))
+  mina_containers = list(filter(lambda c: c['name'] in [ 'coda', 'seed', 'coordinator', 'archive' ], containers))
 
   def restarted_recently(c):
 
@@ -48,7 +48,32 @@ def collect_cluster_crashes(v1, namespace, cluster_crashes):
 
 # ========================================================================
 
-from telemetry_metrics import collect_telemetry_metrics
+def pods_with_no_new_logs(v1, namespace, nodes_with_no_new_logs):
+  print('counting pods with no new logs')
+  pods = v1.list_namespaced_pod(namespace, watch=False)
+
+  ten_minutes = 10 * 60
+
+  count = 0
+  for pod in pods.items:
+    containers = pod.status.container_statuses
+    mina_containers = list(filter(lambda c: c.name in [ 'coda', 'seed', 'coordinator' ], containers))
+    if len(mina_containers) != 0:
+      name = pod.metadata.name
+      recent_logs = v1.read_namespaced_pod_log(name=name, namespace=namespace, since_seconds=ten_minutes)
+      if len(recent_logs) == 0:
+        count += 1
+
+  total_count = len(pods.items)
+
+  fraction_no_new_logs = float(count) / float(total_count)
+  print(count, 'of', total_count, 'pods have no logs in the last 10 minutes')
+
+  nodes_with_no_new_logs.set(fraction_no_new_logs)
+
+# ========================================================================
+
+from node_status_metrics import collect_node_status_metrics
 
 # ========================================================================
 
@@ -88,4 +113,3 @@ def check_seed_list_up(v1, namespace, seeds_reachable):
   seeds_reachable.set(fraction_up)
 
 # ========================================================================
-