+++ src/app/rosetta/lib/construction.ml
@@ -1,6 +1,9 @@
 open Core_kernel
 open Async
 open Rosetta_lib
+
+(* Rosetta_models.Currency shadows our Currency so we "save" it as MinaCurrency first *)
+module MinaCurrency = Currency
 open Rosetta_models
 module Signature = Mina_base.Signature
 module Transaction = Rosetta_lib.Transaction
@@ -290,9 +293,10 @@ module Metadata = struct
           Metadata_data.create ~sender:options.Options.sender
             ~token_id:options.Options.token_id ~nonce
           |> Metadata_data.to_yojson
-          (* TODO: Set this to our default fee, assuming it is fixed when we launch *)
-      ; suggested_fee= [Amount_of.coda (Unsigned.UInt64.of_int 2_000_000_000)]
-      }
+      ; suggested_fee=
+          [ Amount_of.coda
+              (MinaCurrency.Fee.to_uint64
+                 Mina_compile_config.default_transaction_fee) ] }
   end
 
   module Real = Impl (Deferred.Result)
@@ -825,7 +829,7 @@ module Submit = struct
   module Mock = Impl (Result)
 end
 
-let router ~graphql_uri ~logger (route : string list) body =
+let router ~get_graphql_uri_or_error ~logger (route : string list) body =
   [%log debug] "Handling /construction/ $route"
     ~metadata:[("route", `List (List.map route ~f:(fun s -> `String s)))] ;
   let open Deferred.Result.Let_syntax in
@@ -856,6 +860,7 @@ let router ~graphql_uri ~logger (route : string list) body =
         @@ Construction_metadata_request.of_yojson body
         |> Errors.Lift.wrap
       in
+      let%bind graphql_uri = get_graphql_uri_or_error () in
       let%map res =
         Metadata.Real.handle ~env:(Metadata.Env.real ~graphql_uri) req
         |> Errors.Lift.wrap
@@ -887,6 +892,7 @@ let router ~graphql_uri ~logger (route : string list) body =
         @@ Construction_parse_request.of_yojson body
         |> Errors.Lift.wrap
       in
+      let%bind graphql_uri = get_graphql_uri_or_error () in
       let%map res =
         Parse.Real.handle ~env:(Parse.Env.real ~graphql_uri) req
         |> Errors.Lift.wrap
@@ -908,6 +914,7 @@ let router ~graphql_uri ~logger (route : string list) body =
         @@ Construction_submit_request.of_yojson body
         |> Errors.Lift.wrap
       in
+      let%bind graphql_uri = get_graphql_uri_or_error () in
       let%map res =
         Submit.Real.handle ~env:(Submit.Env.real ~graphql_uri) req
         |> Errors.Lift.wrap