+++ helm/block-producer/CHANGELOG.md
@@ -11,3 +11,14 @@
 ------
 - Add static peer id support
 
+0.5.1:
+------
+- add peers list URL chart value
+
+1.0.0:
+------
+- coda -> mina
+
+1.0.1:
+------
+- enable daemon blockchain archival capability